requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==3.13.0
djangorestframework==3.15.2
psycopg2-binary==2.9.9
sqlparse==0.5.1
typing_extensions==4.12.2
celery==5.4.0
redis==5.0.0
gunicorn==20.1.0
django-environ==0.9.0
django-filter==23.2

